requirements.txt
flask~=2.2.3
gunicorn
python-binance